pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pg-essays-epub"
version = "1.0.0"
description = "A tool to scrape Paul Graham's essays and compile them into a Kobo-friendly EPUB"
authors = [{name = "Tom Yan"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "ebooklib>=0.18",
    "lxml>=4.9.0",
    "Pillow>=10.0.0",
]

[project.scripts]
pg-to-epub = "pg_to_epub:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
]

